refactor(JobsServerApi): add Ports and implements custom errors

+++ backend/src/domain/use-cases/jobs/delete-cron.ts
@@ -1,8 +1,8 @@
-import { request } from "http";
-import { Either, right } from "../../../shared/Either";
+import { Either, left, right } from "../../../shared/Either";
 import { ScheduleRepositoryProtocol } from "../_ports/repositories/background-jobs-repository";
+import { NotExistsError } from "../errors/notFound-error";
 
-export class DeleteCron {
+export class DeleteCron implements DeleteCronUseCaseProtocol.UseCase {
   private readonly repository: ScheduleRepositoryProtocol;
 
   constructor(repository: ScheduleRepositoryProtocol) {
@@ -12,6 +12,14 @@ export class DeleteCron {
   async execute(
     request: DeleteCronUseCaseProtocol.Request
   ): DeleteCronUseCaseProtocol.Response {
+    const exists = await this.repository.getScheduleByQueue({
+      Queue: request.Name,
+    });
+
+    if (exists == null) {
+      return left(new NotExistsError(`Queue name ${request.Name} not exists`));
+    }
+
     const data = await this.repository.deleteSchedule({
       Name: request.Name,
     });
@@ -25,4 +33,8 @@ export namespace DeleteCronUseCaseProtocol {
     Name: string;
   };
   export type Response = Promise<Either<Error, any | null>>;
+
+  export interface UseCase {
+    execute(request: Request): DeleteCronUseCaseProtocol.Response;
+  }
 }
